Quita los warnings

+++ Practica1/plantillaP1/src/HomedFrog.cpp
@@ -13,10 +13,10 @@ void HomedFrog::render() const {
 }
 
 Collision HomedFrog::checkCollision(const SDL_FRect& other) const {
-	int width = m_pTexture->getFrameWidth();
-	int height = m_pTexture->getFrameHeight();
-	int x = int(m_Pos.getX());
-	int y = int(m_Pos.getY());
+    float width = (float)m_pTexture->getFrameWidth();
+    float height = (float)m_pTexture->getFrameHeight();
+    float x = (float)m_Pos.getX();
+    float y = (float)m_Pos.getY();
 
 	SDL_FRect rect = { x, y, width, height };
 	Collision ret;
